+++ src/org/sosy_lab/cpachecker/cfa/ast/AAstNode.java
@@ -11,6 +11,7 @@ package org.sosy_lab.cpachecker.cfa.ast;
 import java.io.Serializable;
 import org.sosy_lab.cpachecker.cfa.ast.c.CAstNodeVisitor;
 import org.sosy_lab.cpachecker.cfa.ast.java.JAstNodeVisitor;
+import org.sosy_lab.cpachecker.cfa.ast.svlib.SvLibAstNodeVisitor;
 
 public interface AAstNode extends Serializable {
 
@@ -60,8 +61,10 @@ public interface AAstNode extends Serializable {
           R,
           R1 extends R,
           R2 extends R,
+          R3 extends R,
           X1 extends Exception,
           X2 extends Exception,
-          V extends CAstNodeVisitor<R1, X1> & JAstNodeVisitor<R2, X2>>
-      R accept_(V v) throws X1, X2;
+          X3 extends Exception,
+          V extends CAstNodeVisitor<R1, X1> & JAstNodeVisitor<R2, X2> & SvLibAstNodeVisitor<R3, X3>>
+      R accept_(V v) throws X1, X2, X3;
 }