+++ src/org/sosy_lab/cpachecker/util/predicates/pathformula/pointeraliasing/CExpressionVisitorWithPointerAliasing.java
@@ -911,9 +911,7 @@ class CExpressionVisitorWithPointerAliasing
     BooleanFormula condition =
         conv.fmgr.makeEqual(asValueFormula(parameterAtIndex, CNumericTypes.CHAR), nullTerminator);
     return conv.bfmgr.ifThenElse(
-        condition,
-        conv.fmgr.makeNumber(conv.getFormulaTypeFromCType(returnType), index),
-        otherwise);
+        condition, conv.fmgr.makeNumber(conv.getFormulaTypeFromType(returnType), index), otherwise);
   }
 
   /**